+++ src/Plugin/AnonymizeUserId.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Omikron\Factfinder\Plugin;
+
+use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Store\Model\ScopeInterface;
+use Magento\Store\Model\ScopeInterface as Scope;
+use Omikron\Factfinder\Model\SessionData;
+
+class AnonymizeUserId
+{
+    private const PATH_ANONYMIZE_USER_ID   = 'factfinder/advanced/anonymize_user_id';
+
+    private $scopeConfig;
+
+    public function __construct(ScopeConfigInterface $scopeConfig)
+    {
+        $this->scopeConfig = $scopeConfig;
+    }
+
+    /**
+     * @param SessionData $_
+     * @param int         $userId
+     *
+     * @return string|int
+     * @SuppressWarnings(PHPMD)
+     */
+    public function afterGetUserId(SessionData $_, string $userId)
+    {
+        return $this->isAnonymizationEnabled() ? md5($userId) : $userId; // phpcs:ignore
+    }
+
+    private function isAnonymizationEnabled(): bool
+    {
+        return $this->scopeConfig->isSetFlag(self::PATH_ANONYMIZE_USER_ID, ScopeInterface::SCOPE_STORES);
+    }
+}