+++ src/Plugin/AnonymizeUserId.php
@@ -29,7 +29,7 @@ class AnonymizeUserId
      */
     public function afterGetUserId(SessionData $_, string $userId)
     {
-        return $this->isAnonymizationEnabled() ? md5($userId) : $userId; // phpcs:ignore
+        return $this->isAnonymizationEnabled() && $userId !== '' ? md5($userId) : $userId; // phpcs:ignore
     }
 
     private function isAnonymizationEnabled(): bool