+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ### Change
 - upgrade of package versions
 
+### Add
+ - Tracking
+   - add `user-id` to all kind of tracking if user is logged in
+
 ## [v3.6.0] - 2022.08.25
 ### BREAKING
  - `Omikron\Factfinder\Model\Export\Catalog\Products`